Re-organize color chooser views
